requirements.txt
# Install ONE of these two blocks depending on which free API you use:

# ── Option A: Google Gemini (recommended) ──────────────────────────
# Free at https://aistudio.google.com/app/apikey
google-generativeai>=0.5.0

# ── Option B: Groq ─────────────────────────────────────────────────
# Free at https://console.groq.com
# Uncomment the line below and comment out google-generativeai above
# groq>=0.9.0

# ── Always required ────────────────────────────────────────────────
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.7.0